requirements.txt
torch~=1.10.0
numpy~=1.21.4
pycocotools~=2.0.3
torchvision~=0.11.1
Pillow~=8.4.0
opencv-python~=4.5.3.56
matplotlib~=3.5.0
pandas~=1.3.4
gdown